修复带小数的数字型值导出Excel时会多出几位小数的BUG

+++ ureport2-core/src/main/java/com/bstek/ureport/export/excel/high/builder/ExcelBuilderDirect.java
@@ -153,7 +153,7 @@ public class ExcelBuilderDirect extends ExcelBuilder {
 		        			cell.setCellType(CellType.STRING);
 		        		}else if(obj instanceof Number){
 		        			BigDecimal bigDecimal=Utils.toBigDecimal(obj);
-		        			cell.setCellValue(bigDecimal.floatValue());
+		        			cell.setCellValue(bigDecimal.doubleValue());
 		        			cell.setCellType(CellType.NUMERIC);
 		        		}else if(obj instanceof Boolean){
 		        			cell.setCellValue((Boolean)obj);
